Remove torch distributed from CODEOWNERS (#112813)
After adding support for labeler, we don't need CODEOWNERS.

This change will cause the distributed team members previously listed in
CODEOWNERS to stop being auto-added as reviewers on PRs touching these
files.  The preceding PR adds labeler support for these same sets of
files, and contains instructions for adding yourself to be cc'd for that
label.

It is preferable to be auto-cc'd rather than auto-tagged as reviewer, so
that there is more signal in the reviewers list (either someone opted in
which shows the PR author someone is likely looking at it, or the PR
author added someone specifically which is a stronger notification to
the tagged reviewer than the blanket CODEOWNERS behavior.

Pull Request resolved: https://github.com/pytorch/pytorch/pull/112813
Approved by: https://github.com/wanchaol, https://github.com/fduwjj