Rewrite docs so that it is OK to use record_stream before uses (#113282)
The previous documentation did not appear to accurately describe
the actual semantics in CUDA caching allocator.

When you record stream, we only record a stream use:

```
  void recordStream(Block* block, cuda::CUDAStream stream) {
    std::lock_guard<std::recursive_mutex> lock(mutex);
    if (stream.stream() == block->stream) {
      // ignore uses on the allocation stream, since those don't require any
      // special synchronization
      return;
    }
    block->stream_uses.insert(stream);
  }
```

It is only at deallocation time when we actually install an event on
stream uses that we will subsequently query to determine if the block
can be reused or not.

Signed-off-by: Edward Z. Yang <ezyang@meta.com>
Pull Request resolved: https://github.com/pytorch/pytorch/pull/113282
Approved by: https://github.com/Skylion007, https://github.com/albanD